Add CSS, syntax highlighting, more content
I've added CSS to work with the HTML Erlang code generated using
erl_to_html: it's not a turn-key solution as it requires some
formatting for comments, newlines, blank lines, etc.

Added a section on running Erlang from a file. This was a major
complaint that I heard right away.